Lab 5 non-shared kernel stack
We allocate kernel stacks dynamically when a new task is created using
`page_alloc` & `page_insert` in `setupkstack`. Contents of new stack
have to be the same as the old one, so `copykstack` copy the stack
content to the newly assigned address space. Finally `removekstack`
frees kernel the corresponding kernel stack when freeing a task.